Le type de PI finalement c'est un double !

+++ outils.h
@@ -1,6 +1,7 @@
 #ifndef __OUTILS_H__
 #define __OUTILS_H__
 
-const int PI = 3.141259999999;
+const double PI = 3.141259999999;
+
 
 #endif
